+++ src/main/java/com/udxp/service/CategoryService.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 
 public interface CategoryService {
     CategoryResponse createCategory(CategoryCreateRequest request);
-    CategoryResponse updateCategory(int id, CategoryCreateRequest request);
+    CategoryResponse updateCategory(Long id, CategoryCreateRequest request);
     void deleteCategory(String categoryName);
     Page<String> getCategoryNames(Pageable pageable);
 }
